refactor(server): remove sync v2 + drop sync tables + cursor columns (phase 1/3)
Sync v2 (offline-first delta protocol) is fully removed:

- Collab module no longer wires /api/sync/v2/*. ShareService + Yjs
  WebSocket collab remain. Stale "SQLite" comment in collab/index.ts
  corrected to Postgres.
- Sync v2 writer call sites stripped from notes/services/{note,folder,
  tag}.service.ts, notes/services/backfill/tags-backfill.ts,
  notes/routes/trash.routes.ts, and knowledge/services/cursor.service.ts
  (deleted).
- Schema: db/schema/sync.ts renamed to collab.ts (yjsDocument +
  yjsUpdate only). SyncDeletion + SyncCursor tables dropped. The dead
  cursor: bigint columns dropped from every domain table that had one
  (Folder, GraphEdge, NoteTag, NoteVersion, Tag, TrashItem, NoteShare,
  Settings, InstalledPlugin).
- Share serialization no longer exposes a cursor field; shareSchema
  trimmed accordingly.
- Test helpers no longer truncate SyncDeletion/SyncCursor; collab
  __tests__/shares.test.ts no longer asserts sync v2 401s.
- OpenAPI snapshot + SDK types regenerated.

Phase 2 (delete @azrtydxb/core + core-react) and Phase 3 (docs
sweep) follow.
