+++ packages/server/src/modules/notes/services/folder.service.ts
@@ -1,30 +1,11 @@
 import type { FastifyInstance } from "fastify";
-import { and, asc, eq, sql } from "drizzle-orm";
+import { and, asc, eq } from "drizzle-orm";
 import { folder } from "../../../db/schema/notes.js";
-import { syncCursor } from "../../../db/schema/sync.js";
-
-/**
- * Increment the per-user sync cursor and return the new value.
- * Local copy of the legacy cursor service — kept inside the notes module so it
- * has no module-level Prisma singleton.
- */
-async function incrementCursor(app: FastifyInstance, userId: string): Promise<bigint> {
-  const [row] = await app.db
-    .insert(syncCursor)
-    .values({ userId, cursor: 1n })
-    .onConflictDoUpdate({
-      target: syncCursor.userId,
-      set: { cursor: sql`${syncCursor.cursor} + 1` },
-    })
-    .returning({ cursor: syncCursor.cursor });
-  return row.cursor;
-}
 
 export class FolderService {
   constructor(private readonly app: FastifyInstance) {}
 
   async create(userId: string, input: { path: string; parentPath?: string }) {
-    const cursor = await incrementCursor(this.app, userId);
     const parent = input.parentPath
       ? await this.app.db.query.folder.findFirst({
           where: and(eq(folder.userId, userId), eq(folder.path, input.parentPath)),
@@ -37,7 +18,6 @@ export class FolderService {
         path: input.path,
         parentId: parent?.id ?? null,
         version: 1,
-        cursor,
       })
       .returning();
     return created;
@@ -65,5 +45,3 @@ export class FolderService {
     return deleted;
   }
 }
-
-export { incrementCursor };
